make cacheable a generic store type
